feat: add faq fng about and faq fng food commands
